fix push wrong at challenge-08
